func_app/requirements.txt
azure-functions
azure-identity
azure-communication-email
requests
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
